# api/openapi/valori-v1.yaml
components:
  schemas:
    ApiError:
      description: |-
        The canonical error body, as a schema-bearing DTO.

        [`valori_engine::EngineError`] produces this shape at runtime but lives in
        a crate that does not depend on `utoipa`, so the schema is declared here —
        the translation layer §36 asks for, rather than a re-export of an internal
        type.
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          description: Stable machine-readable code. Branch on this.
        error:
          description: Human-readable message. Do not parse.
          type: string
      required:
      - error
      - code
      type: object
...
    ErrorCode:
      description: |-
        Mirror of [`valori_engine::ErrorCode`] for schema generation.

        `tests/api_contract.rs` diffs the runtime enum against the committed YAML;
        this type exists so the generated document carries the same closed set.
      enum:
      - validation_error
      - unauthorized
      - forbidden
      - not_found
      - collection_not_found
      - record_not_found
      - dimension_mismatch
      - invalid_metric
      - invalid_index
      - index_build_failed
      - conflict
      - capacity_exceeded
      - not_leader
      - unavailable
      - not_implemented
      - internal_error
      type: string
...
    IngestAcceptedResponse:
      description: |-
        The `202` body of `POST /v1/ingest` when `async: true`.

        The async branch has always returned this object; the contract used to
        declare the `202` with no content at all, so a generated client saw
        `never` and had no typed way to reach `job_id` — the one field the whole
        async flow depends on, since it is what `GET /v1/ingest/status/{job_id}`
        takes.
      properties:
        collection:
          type: string
        job_id:
          description: Poll `GET /v1/ingest/status/{job_id}` with this id.
          type: string
        ok:
          type: boolean
        status:
          description: Always `processing` on this response.
          type: string
      required:
      - ok
      - job_id
      - status
      - collection
      type: object
    IngestChunk:
      description: One chunk produced by a chunking strategy.
      properties:
        index:
          description: 0-based position in the chunk sequence.
          minimum: 0
          type: integer
        text:
          description: Full chunk text ready to embed.
          type: string
        title:
          description: Section title (tree strategy) or empty string.
          type: string
      required:
      - index
      - title
      - text
      type: object
    IngestDocumentRequest:
      properties:
        chunk_overlap:
          description: Fixed-strategy overlap in chars (default 200).
          minimum: 0
          type:
          - integer
          - 'null'
        chunk_size:
          description: Fixed-strategy chunk size in chars (default 1000).
          minimum: 0
          type:
          - integer
          - 'null'
        collection:
          description: Collection to ingest into (default = "default").
          type:
          - string
          - 'null'
        source:
          description: Source label stored in metadata (e.g. filename). Optional.
          type:
          - string
          - 'null'
        strategy:
          description: 'Chunking strategy: `auto` | `tree` | `conversation` | `sentence` | `fixed`.'
          type:
          - string
          - 'null'
        text:
          description: Raw text content of the document.
          type: string
      required:
      - text
      type: object
    IngestDocumentResponse:
      properties:
        chunk_count:
          description: Total number of chunks produced.
          minimum: 0
          type: integer
        chunks:
          description: |-
            The chunks. Caller embeds each `text`, inserts the vector, records
            `record_id` → chunk for provenance.
          items:
            $ref: '#/components/schemas/IngestChunk'
          type: array
        collection:
          description: Collection the document was targeted at.
          type: string
        strategy_used:
          description: Strategy that was actually used (useful when `strategy="auto"`).
          type: string
      required:
      - strategy_used
      - chunk_count
      - chunks
      - collection
      type: object
...
    IngestRequest:
      properties:
        async:
          type:
          - boolean
          - 'null'
        chunk_overlap:
          minimum: 0
          type:
          - integer
          - 'null'
        chunk_size:
          minimum: 0
          type:
          - integer
          - 'null'
        collection:
          type:
          - string
          - 'null'
        source:
          type:
          - string
          - 'null'
        strategy:
          type:
          - string
          - 'null'
        text:
          type: string
      required:
      - text
      type: object
    IngestResponse:
      properties:
        chunk_count:
          minimum: 0
          type: integer
        collection:
          type: string
        document_node_id:
          format: int32
          minimum: 0
          type: integer
        ok:
          type: boolean
        operation_id:
          description: |-
            Fetch `GET /v1/operations/:id/execution` with this id for the full
            per-stage execution breakdown (Execution Explorer).
          type: string
        record_ids:
          items:
            format: int32
            minimum: 0
            type: integer
          type: array
        strategy_used:
          type: string
      required:
      - ok
      - document_node_id
      - strategy_used
      - chunk_count
      - record_ids
      - collection
      - operation_id
      type: object
...
  securitySchemes:
    BearerAuth:
      description: Project API key or the legacy VALORI_AUTH_TOKEN bearer token.
      scheme: bearer
      type: http
info:
  description: 'The Valori Data Plane REST API v1, emitted directly from `#[utoipa::path]` annotations on the registered axum handlers. Every public route the Rust routers register appears here and nothing else does — `scripts/verify-api-route-contract.py` proves that equality on every run of the contract gate, and fails on any drift in either direction. Administrative and operator-internal routes (key management, cluster membership, replication streams, `/metrics`) are deliberately excluded: they are served by the node but are not part of the SDK surface.'
  license:
    identifier: MIT OR Apache-2.0
    name: MIT OR Apache-2.0
  title: Valori Data Plane API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/ingest:
    post:
      description: 'The full pipeline in one call. Requires `VALORI_EMBED_PROVIDER`. With `async: true` the call returns immediately and progress is polled through `GET /v1/ingest/status/{job_id}`.'
      operationId: ingest_document
      parameters:
      - in: query
        name: async
        required: false
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
          description: Document ingested
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestAcceptedResponse'
          description: Accepted for background processing; poll `GET /v1/ingest/status/{job_id}`
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Empty text, or the chunker produced no chunks
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'Missing or invalid credentials. The auth middleware rejects the request before the handler runs; `attach_error_code` renders the rejection as a canonical `ApiError` with `code: "unauthorized"`.'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'The presented key authenticated but its scope does not satisfy this operation''s `x-required-scope`. Rendered as a canonical `ApiError` with `code: "forbidden"`.'
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Text exceeds the maximum ingest size
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: VALORI_EMBED_PROVIDER is not configured
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: The ingest pipeline failed
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: The embedding provider failed
      security:
      - BearerAuth: []
      summary: Chunk, embed, and insert a document
      tags:
      - ingest
      x-required-scope: read_write
      x-sdk: true
  /v1/ingest/document:
    post:
      description: 'Stateless: no embedding provider is called and nothing is written. Use it to preview how a strategy will split a document before committing to an ingest.'
      operationId: chunk_document
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestDocumentRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestDocumentResponse'
          description: Chunks, with the strategy that produced them
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Empty text, or text over the size limit
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'Missing or invalid credentials. The auth middleware rejects the request before the handler runs; `attach_error_code` renders the rejection as a canonical `ApiError` with `code: "unauthorized"`.'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'The presented key authenticated but its scope does not satisfy this operation''s `x-required-scope`. Rendered as a canonical `ApiError` with `code: "forbidden"`.'
      security:
      - BearerAuth: []
      summary: Chunk a document without embedding or storing it
      tags:
      - ingest
      x-required-scope: read_write
      x-sdk: true
...
servers:
- description: 'The node this document was generated by. Valori is self-hosted: point your client at your own node''s base URL.'
  url: /